==> scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Instructions on how to acquire the data from U.S. GSS
# Author: David James Dimalanta, Harrison Huang, Michael Fang
# Date: 8 March 2024
# Contact: david.dimalanta@mail.utoronto.ca
# License: MIT

#### Download data ####

# Due to licensing and copyright restrictions, we are unable to provide a direct link to download the data from the General Social Survey (GSS) but we will provide steps on how to aquire the date yourself

# First, visit the GSS website at https://gss.norc.org/Get-The-Data
# Then, to download the GSS data, you may be required to register for an account.
# Once registered, you are able to view different variables withing the GSS Data
# Please see the varaibles we used below:

#Year
#Wrkstat: labor force status
#Sex: respondents sex
#Educ: highest year of school completed
#Fejobaff: for or against preferential hiring of women 
#Fehire: Should hire and promote women 
#Fefam: better for man to work, woman tend home 
#Discaffw: a woman won't get a job or promotion 

# You can then use these variables to then tabulate the data to create datasets that can be used for the next steps
